Modularized slack tests for Python + Java
